Split into `base-docker` and `base-action` images
This is because we want to use a standard base-docker image in
a non-action context, so split out the action entrypoint stuff into
a separate image.

Additionally, scheduled a clean build every week. This ensure that we'll
have an up to date base image.

However, it does mean that our users will need to pull down entire
images when they are rebuilt, as the base layer will have changed. So
it's important they are as small as possible.

Also made a small tweak to the app installer script, to be more
efficient when passed a list of packages as cli args.