fix the build arg based labels
Due to quirk of multi-stagge docker builds, you need to set the build
args when you build all targets of the Dockerfile.

We do this by setting the args for the whole Makefile, and building the
whole lot in one invocation. That should mean that all images have the
same sha/build time labels